% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.r
\name{arm_bin_combn}
\alias{arm_bin_combn}
\title{Create Retrospective Response Combinations}
\usage{
arm_bin_combn(arm, resps)
}
\arguments{
\item{arm}{the enrollment arm data.frame.}

\item{resps}{the number of responses.}
}
\description{
Create Retrospective Response Combinations
}
