% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.r
\name{arm_bin_resample}
\alias{arm_bin_resample}
\title{Create Retrorepective Binary Response Samples}
\usage{
arm_bin_resample(arm, resps, size = 1)
}
\arguments{
\item{arm}{the enrollment arm data.frame.}

\item{resps}{the number of responses.}

\item{size}{the number of trials to sample. Default 1.}
}
\description{
Create Retrorepective Binary Response Samples
}
