% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arm.r
\name{arm_enroll}
\alias{arm_enroll}
\title{Enroll an Arm for a Specified Number of Patients}
\usage{
arm_enroll(size, sampler = function() 1, init_period = 0L)
}
\arguments{
\item{size}{the enrollment size of the arm}

\item{sampler}{function to sample enrollment. Default 1 enrollment per 
period.}

\item{init_period}{the initial period. Default 0.}
}
\description{
Enroll an Arm for a Specified Number of Patients
}
