% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05arrivalTimeDiscrete.R
\name{arrivalTimeDiscrete}
\alias{arrivalTimeDiscrete}
\title{arrivalTimeDiscrete}
\usage{
arrivalTimeDiscrete(data, t0 = 0)
}
\arguments{
\item{data}{a dataframe containing a time and a rate column}

\item{t0}{the time of the last arrival}
}
\description{
A function for generating event times from a stepwise rate function. For event times leave t0 at the default (0). For arrival times set t0 to the last arrival time.
}
\examples{
data = data.frame(time = c(0,1,2,3,4,5),rate = c(0.5,2,3,4,6,7))
t0 = 0
arrivalTime(data, 0)
hist(sapply(1:10000, function(x) arrivalTime(data, 0)))
}

