% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial.r
\name{bin_trial_resample}
\alias{bin_trial_resample}
\title{Resample Trial Enrollment and Binary Outcomes}
\usage{
bin_trial_resample(resps, size, name, num_samples,
  sampler = lapply(seq_along(resps), function(x) function() 1))
}
\arguments{
\item{resps}{the number of responses per arm.}

\item{size}{the size of each arm.}

\item{name}{the name of each arm.}

\item{num_samples}{the number of resamples to draw.}

\item{sampler}{function to sample enrollment. Default 1 enrollment per
period.}
}
\description{
Resample Trial Enrollment and Binary Outcomes
}
