% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01compare.R
\name{compare}
\alias{compare}
\title{compare}
\usage{
compare(outFolder, p = NULL, fun = NULL, colNames = NULL, ...)
}
\arguments{
\item{outFolder}{The folder with all the .Rdata files containing results objects}

\item{p}{(optional) A list of terms which will be found in the results object. This acts as a filter only keeping those results object which match.}

\item{fun}{A function to generate a row of the summary data.frame. This function should take at least two arguments: object and file.name.}

\item{colNames}{The column names to use for the /code{data.frame}.}

\item{...}{Additional arguments to pass to fun.}
}
\description{
A function to generate summary comparisons of multiple simulations. This is made reasonably flexible by allowing the user to provide a function to extract the information they require from each results object.
}

