% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00classes.R
\docType{class}
\name{decision-class}
\alias{decision}
\alias{decision-class}
\title{decision}
\description{
This class is used to determine what is going on in the trial at any given time. It is returned from the /code{/link{readyForAnalysis}} and /code{/link{makeDecisions}} functions within the \code{\link{trialDesign}} class.
}
\section{Slots}{

\describe{
\item{\code{analyse}}{TRUE/FALSE whether the trials should fit the model}

\item{\code{continue}}{TRUE/FALSE whether the trial should continue}

\item{\code{cohort}}{A number specifying the current cohort}

\item{\code{recruiting}}{TRUE/FALSE whether new patients can be recruited to the trial at present}

\item{\code{stopReason}}{A string detailing why the trial is stopping}

\item{\code{cohortSize}}{The number of patients to recruit before running the /code{/link{readyForAnalysis}} function}
}}

