% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00classes.R
\docType{class}
\name{experimentalModel-class}
\alias{experimentalModel}
\alias{experimentalModel-class}
\title{experimentalModel}
\description{
A class and generator function to store a function which runs the model and returns detailed results of the run. The more results it returns the easier it will be to make different decisions. If this is purely a rule based approach then the default for fun can be used which returns NULL. Then the rules can be written entirely in the \code{\link{makeDecisions}} component.
}
\section{Slots}{

\describe{
\item{\code{fun}}{Fits the model. This function should allow any number of arguments to be passed to it using \dots. This function can return anything but the more useful information the better.

To reduce the risk of errors we recommend specifying the inputs required without defaults. This way if none are specified in p for the \code{\link{trialDesign}} class then it will throw a meaningful error.}
}}

