% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01genericMethods.R
\name{fitModel}
\alias{fitModel}
\title{fitModel}
\usage{
fitModel(object, ...)
}
\description{
Fit the model to the data and return the model
}
\section{Slots}{

\describe{
\item{\code{object}}{An object with parent class /code{/link{trialDesign}}}
}}

