% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.r
\name{fixed_duration_sampler}
\alias{fixed_duration_sampler}
\title{Sample Enrollment over a Fixed Period}
\usage{
fixed_duration_sampler(size, max_duration = max(size),
  multi_enroll_period = TRUE)
}
\arguments{
\item{size}{the number of enrollees per arm.}

\item{max_duration}{the maximum of the trial. Default 'max(size)'.}

\item{multi_enroll_period}{can there be multiple enrollees in a period?
Default TRUE.}
}
\description{
Sample Enrollment over a Fixed Period
}
