% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01genericMethods.R
\name{getBaseline}
\alias{getBaseline}
\title{getBaseline}
\usage{
getBaseline(object, ...)
}
\description{
Creates baseline data for the next patient. The data to add to should be stored in object@data
}
\section{Slots}{

\describe{
\item{\code{object}}{An object with parent class \code{\link{trialDesign}}}
}}
\seealso{
\code{\link{simulateBaseline}}

\code{\link{getData}}, \code{\link{getBaseline}}, \code{\link{getOutcome}}, \code{\link{getTrigger}} \code{\link{getmodel}}, \code{\link{getDecision}}
}

