% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01genericMethods.R
\name{getDecision}
\alias{getDecision}
\title{getDecision}
\usage{
getDecision(object, ...)
}
\description{
Fit the model to the data and return the decision
}
\section{Slots}{

\describe{
\item{\code{object}}{An object with parent class /code{/link{trialDesign}}}
}}

