% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01genericMethods.R
\name{getOutcome}
\alias{getOutcome}
\title{getOutcome}
\usage{
getOutcome(object, ...)
}
\description{
Creates outcome data for the next patient. The data to add to should be stored in object@data. The decision in object@p$decision is used to determine any special requirements such as inclusion due to biomarkers or trial recruitment paused.
}
\section{Slots}{

\describe{
\item{\code{object}}{An object with parent class \code{\link{trialDesign}}}
}}
\seealso{
\code{\link{simulateOutcome}}

\code{\link{getData}}, \code{\link{getBaseline}}, \code{\link{getOutcome}}, \code{\link{getTrigger}} \code{\link{getmodel}}, \code{\link{getDecision}}
}

