% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01genericMethods.R
\name{getTrigger}
\alias{getTrigger}
\title{getTrigger}
\usage{
getTrigger(object, ...)
}
\description{
Runs the object@triggerAnalysis creating a decision class object to determine if the analysis should be run. The data to add to should be stored in object@data.
}
\section{Slots}{

\describe{
\item{\code{object}}{An object with parent class \code{\link{trialDesign}}}
}}
\seealso{
\code{\link{readyForAnalysis}}

\code{\link{getData}}, \code{\link{getBaseline}}, \code{\link{getOutcome}}, \code{\link{getTrigger}} \code{\link{getmodel}}, \code{\link{getDecision}}
}

