% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00classes.R
\docType{class}
\name{makeDecisions-class}
\alias{makeDecisions}
\alias{makeDecisions-class}
\title{makeDecisions}
\description{
Takes the data and the model and makes decisions as to what should happend next. Returns an object of class /code{/link{decision}}
}
\section{Slots}{

\describe{
\item{\code{fun}}{Makes decisions based on the data and fitted model. This function should allow any number of arguments to be passed to it using \dots. This function can return anything but the more useful information the better.

To reduce the risk of errors we recommend specifying the inputs required without defaults. This way if none are specified in p for the \code{\link{trialDesign}} class then it will throw a meaningful error.}
}}

