% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02parallelTrial.R
\name{parallelTrial}
\alias{parallelTrial}
\title{parallelTrial}
\usage{
parallelTrial(fun, nsim, vars, parallel)
}
\arguments{
\item{fun}{The function to run nsim times}

\item{nsim}{The number of simulations to run}

\item{vars}{The variables which need to be passes to each cluster}

\item{parallel}{A logical value to determine if the runs should be made parallel}
}
\value{
This functions returns a list of results for fun. This is identical to a call to \code{\link{lapply}}.
}
\description{
This is a utility function for running simulations in parallel.
}

