% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampler.r
\name{poisson_sampler}
\alias{poisson_sampler}
\title{Sample Enrollment from a Poisson Distribution}
\usage{
poisson_sampler(lambda)
}
\arguments{
\item{lambda}{the poisson rate parameter (one per arm).}
}
\description{
Sample Enrollment from a Poisson Distribution
}
