% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00classes.R
\docType{class}
\name{readyForAnalysis-class}
\alias{readyForAnalysis}
\alias{readyForAnalysis-class}
\title{readyForAnalysis}
\description{
Decides if it is time to run the model. Take the data and an object of /code{/link{Decision}} class and returns a /code{/link{Decision}} class object.
}
\section{Slots}{

\describe{
\item{\code{fun}}{Decides if it is time to run the model. This function should allow any number of arguments to be passed to it using \dots. This function must return a /code{/link{Decision}} class object.

To reduce the risk of errors we recommend specifying the inputs required without defaults. This way if none are specified in p for the \code{\link{trialDesign}} class then it will throw a meaningful error.}
}}

