% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00classes.R
\docType{class}
\name{setupFunction-class}
\alias{setupFunction}
\alias{setupFunction-class}
\title{setupFunction}
\description{
Provides a place to do random setup for each run. Particularly useful for non patient level random variables such as surgeon effect. May also be used to compute things like assurance given a prior belief and coverage given a distribution for data rather than a known effect.
}
\section{Slots}{

\describe{
\item{\code{fun}}{Does one time random number generation saving results to p for use in the other functions.}
}}

