% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01genericMethods.R
\name{simTrial}
\alias{simTrial}
\title{simTrial}
\usage{
simTrial(object, ...)
}
\description{
Runs trial level simulations
}
\section{Slots}{

\describe{
\item{\code{object}}{An object with parent class \code{\link{trialDesign}}}

\item{\code{parallel}}{TRUE/FALSE Whether to run the simulations in parallel or not}

\item{\code{seed}}{Set the seed at the start of the run}
}}

