% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00classes.R
\docType{class}
\name{simulateBaseline-class}
\alias{simulateBaseline}
\alias{simulateBaseline-class}
\title{simulateBaseline}
\description{
Simulates baseline data one patient at a time and appends it to the current data.
}
\section{Slots}{

\describe{
\item{\code{fun}}{Does the simulation and appends the new data. This function should allow any number of arguments to be passed to it using \dots. This function must return the data.

To reduce the risk of errors we recommend specifying the inputs required without defaults. This way if none are specified in p for the \code{\link{trialDesign-class}} class then it will throw a meaningful error.

This function should also be capable of creating the dataframe if provided with an empty data.frame.}
}}

