% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04Utility_strictNumLength.R
\name{strictNumLength}
\alias{strictNumLength}
\title{strictNumLength}
\usage{
strictNumLength(i, d = 4)
}
\arguments{
\item{i}{the number to make strict}

\item{d}{the length of the output number}
}
\description{
A vectorised utility tool for making numbers the same length. This is usefull for getting filenames in the right order since R thinks stuff_2 comes after stuff_10. Instead this will return numbers such as 0002 and 0010.
}
\examples{

strictNumLength(c(1,2,3))


}

