% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00classes.R
\docType{class}
\name{trialDesign-class}
\alias{trialDesign}
\alias{trialDesign-class}
\title{trialDesign}
\description{
This is the parent class for all trial designs.
}
\section{Slots}{

\describe{
\item{\code{triggerAnalysis}}{An object of class \code{\link{readyForAnalysis}} which makes decisions about when tofit the model and pause recruitment}

\item{\code{model}}{An object of class \code{\link{experimentalModel}} which fits the model and returns a model object}

\item{\code{decision}}{An object of class \code{\link{makeDecisions}} which makes decisions based on the model and data}

\item{\code{data}}{A data.frame of initial data when simulating part way through a trial}

\item{\code{seed}}{An integer used to seed the simulations}

\item{\code{nSim}}{Number of simulations to run}

\item{\code{sims}}{A slot to store a list of the simulated trials run}

\item{\code{p}}{A list of all of the required parameters for the running functions}
}}

