% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01UpdateSimsInFolder.R
\name{UpdateSimsInFolder}
\alias{UpdateSimsInFolder}
\title{UpdateSimsInFolder}
\usage{
UpdateSimsInFolder(inFolder, outFolder, updateFunctions = NULL, ...)
}
\arguments{
\item{inFolder}{Folder of trial designs to run}

\item{outFolder}{Output folder to save results to}

\item{updateFunctions}{An object of the same class as the setup object with updated functions. Allows you to update everything wilh new functions and rerun anything which has changed.}

\item{...}{Arguments passed to simTrial. In particular you may want to set /code{parallel = TRUE}.}
}
\description{
This folder takes a group of designs and runs them one after the other. There is checking mechanisms in order to prevent rerun
}

