% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01genericMethods.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(object, ...)
}
\description{
Creates data for the next patient or patients. The data to add to should be stored in object@data
}
\section{Slots}{

\describe{
\item{\code{object}}{An object with parent class \code{\link{trialDesign}}}
}}

